+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## X.Y.Z (UNRELEASED)
 
 * **New Resource:** `sonatyperepo_task_blobstore_compact` [GH-92]
+* **New Resource:** `sonatyperepo_task_repair_create_browse_nodes` [GH-92]
 * **New Resource:** `sonatyperepo_task_repository_maven_remove_snapshots` [GH-92]
 * **New Resource:** `sonatyperepo_task_repository_docker_gc` [GH-92]
 * **New Resource:** `sonatyperepo_task_repository_docker_upload_purge` [GH-92]